adding input/ output moving coordinate digest func
