+++ ConwaysGameOfLife/Coordinates.cs
@@ -14,22 +14,22 @@ namespace ConwaysGameOfLife
     }
 // should coordinates be a data structure? Should I remove 'behaviour'
 
-    public static List<Coordinate> DigestCoordinates(string coordinatesString)
-    {
-      var coordinatesList = new List<Coordinate>();
-      var coordinates = coordinatesString.Split(" ");
-      foreach (string coord in coordinates)
-      {
-        var rowTryParse = int.TryParse(coord[0].ToString(), out int row);
-        var colTryParse = int.TryParse(coord[2].ToString(), out int column);
+    // public static List<Coordinate> DigestCoordinates(string coordinatesString)
+    // {
+    //   var coordinatesList = new List<Coordinate>();
+    //   var coordinates = coordinatesString.Split(" ");
+    //   foreach (string coord in coordinates)
+    //   {
+    //     var rowTryParse = int.TryParse(coord[0].ToString(), out int row);
+    //     var colTryParse = int.TryParse(coord[2].ToString(), out int column);
 
-        if (rowTryParse && colTryParse)
-        {
-          coordinatesList.Add(new Coordinate (row, column));
-        }
-      }
-      return coordinatesList;
-    }
+    //     if (rowTryParse && colTryParse)
+    //     {
+    //       coordinatesList.Add(new Coordinate (row, column));
+    //     }
+    //   }
+    //   return coordinatesList;
+    // }
 
   }
 }
