+++ ConwaysGameOfLife/IRead.cs
@@ -1,0 +1,8 @@
+namespace ConwaysGameOfLife
+{
+  interface IRead
+  {
+    string ReadInput();
+
+  }
+}
