+++ ConwaysGameOfLife/IWrite.cs
@@ -1,0 +1,7 @@
+namespace ConwaysGameOfLife
+{
+    public interface IWrite
+    {
+        string WriteOutput(); 
+    }
+}
